+++ modules/components/item/RuleGroupExt.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import GroupContainer from "../containers/GroupContainer";
 import Draggable from "../containers/Draggable";
-import {BasicGroup} from "./Group";
+import BasicGroup from "./BasicGroup";
 import {RuleGroupExtActions} from "./RuleGroupExtActions";
 import FieldWrapper from "../rule/FieldWrapper";
 import OperatorWrapper from "../rule/OperatorWrapper";
